Adicionando melhorias

+++ src/models/Tarifa.java
@@ -1,6 +1,6 @@
 package models;
 
 public interface Tarifa {
-    double calcularTarifa(double minutos);
+    double calcularTarifa(long minutos);
    
 }
